Restructure chat handlers to separate files
Greatly reduces chatcommands.js size, groups commands by domain
