+++ lib/commands/ascii.js
@@ -1,89 +1,58 @@
-/*
- * This is useful for quick and dirty custom commands.
- * See chatcommands.js for examples on how to add commands.
- */
-
-import random from 'random';
-
-import {Rank} from './constants.js';
-import {errorLog} from './logger.js';
-import {sendMessagesWithRateLimit, sendPmsWithRateLimit, sleep} from './utils.js';
-
-/** @typedef {import('./cytubebot.js').CytubeBot} CytubeBot */
-
 /**
- * @callback Handler
- * @param {CytubeBot} bot Reference to the CytubeBot.
- * @param {string} username Username of the user that sent the message.
- * @param {string} msg The message to handle, not including the command.
+ * Commands that just print an ASCII into chat.
+ *
+ * They rely on a chat filter in the room replacing %^^% with <br />.
  */
 
-export const /** @type {!Map<string, Handler>} */ HANDLERS = new Map();
+import {RateLimiter} from 'limiter';
 
-const COCK_MAX_LENGTH = 14;
+import {Rank} from '../constants.js';
+import {sleep} from '../utils.js';
 
-HANDLERS.set('cock', (bot, username, msg) => {
-  const target = msg.split(' ')[0] !== '' ? msg.split(' ')[0] : username;
-  const length = Math.round(Math.random() * COCK_MAX_LENGTH);
+/** @typedef {import('./handle.js').Handler} Handler */
 
-  let /** @type {string} */ emote;
-  if (length > 10) {
-    emote = 'gachiHYPER';
-  } else if (length < 4) {
-    emote = 'forsenLaughingAtYou';
-  } else {
-    emote = 'gachiGASM';
-  }
+class Limiters {
+  static FORSEN = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static POT_FRIEND = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static TROLL_DESPAIR = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static DEEZ = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static VALORANT = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static AMONG = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static XQCK = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static PEPE_JAM = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static FUCK_YOU_CHAT = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static AMONG_E = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static GIGACHAD = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static KKONA = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static AYAYA = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static FLIP = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static KAPPA = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static PAG_CHOMP = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static BEDGE = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static STAREGE = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static POGGERS = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static PEEPO_SAD = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static MONKA_S = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static HYPERS = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static YEP = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static MONKA_W = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static PEPE_LAUGH = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static PEPEGA = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static UWU = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static TF = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static GACHI = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static FORGOR = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static FORTNITE = new RateLimiter({tokensPerInterval: 1, interval: 900 * 1_000});
+  static NO_MODS = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static MODS_SMELL_BAD = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static ONION = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
+  static OSU = new RateLimiter({tokensPerInterval: 1, interval: 'minute'});
+}
 
-  bot.sendChatMsg(`${target}'s cock is ${length} inches long ${emote}`);
-});
+export const /** @type {!Map<string, Handler>} */ COMMANDS = new Map();
 
-HANDLERS.set('debugchat', async (bot, username, msg) => {
-  if (username !== 'airforce2700') {
-    bot.sendChatMsg('Only bot guy can run that Okayeg');
-    return;
-  }
-
-  let rows;
-  try {
-    rows = await bot.db.dumpForDebug(msg);
-  } catch (err) {
-    errorLog.log(err);
-    bot.sendChatMsg('Dump failed');
-    return;
-  }
-  if (!rows || rows.length === 0) {
-    bot.sendChatMsg('No data returned');
-    return;
-  }
-
-  await sendMessagesWithRateLimit(bot, rows.map((row) => JSON.stringify(row)));
-});
-
-HANDLERS.set('debugpm', async (bot, username, msg) => {
-  if (username !== 'airforce2700') {
-    bot.sendChatMsg('Only bot guy can run that Okayeg');
-    return;
-  }
-
-  let rows;
-  try {
-    rows = await bot.db.dumpForDebug(msg);
-  } catch (err) {
-    errorLog.log(err);
-    bot.sendChatMsg('Dump failed');
-    return;
-  }
-  if (!rows || rows.length === 0) {
-    bot.sendChatMsg('No data returned');
-    return;
-  }
-
-  await sendPmsWithRateLimit(bot, username, rows.map((row) => JSON.stringify(row)));
-});
-
-HANDLERS.set('forsen', async (bot, username, msg) => {
-  if (!bot.forsenLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('forsen', async (bot, username, msg) => {
+  if (!Limiters.FORSEN.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$forsen is on cooldown');
   }
 
@@ -93,8 +62,8 @@ HANDLERS.set('forsen', async (bot, username, msg) => {
       'forsen3 forsen4 forsenE glizzyL !! !! glizzyR forsenE forsenE forsenE forsenE glizzyL !!');
 });
 
-HANDLERS.set('potfriend', async (bot, username, msg) => {
-  if (!bot.potfriendLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('potfriend', async (bot, username, msg) => {
+  if (!Limiters.POT_FRIEND.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$potfriend is on cooldown');
   }
 
@@ -115,8 +84,8 @@ HANDLERS.set('potfriend', async (bot, username, msg) => {
       '%^^%⠀⠀⠀⠀⠀⠀⠻⢿⡋⠉⠀⠀⠀⠀⠀⠀⠀⠀⠉⣻⡿⠿⠂');
 });
 
-HANDLERS.set('trolldespair', async (bot, username, msg) => {
-  if (!bot.trolldespairLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('trolldespair', async (bot, username, msg) => {
+  if (!Limiters.TROLL_DESPAIR.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$trolldespair is on cooldown');
   }
 
@@ -137,8 +106,8 @@ HANDLERS.set('trolldespair', async (bot, username, msg) => {
       '%^^%⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠉⠉⠉⠉⠈⠄⠄⠄⠄⠄⠄');
 });
 
-HANDLERS.set('deez', async (bot, username, msg) => {
-  if (!bot.deezLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('deez', async (bot, username, msg) => {
+  if (!Limiters.DEEZ.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$deez is on cooldown');
   }
 
@@ -154,8 +123,8 @@ HANDLERS.set('deez', async (bot, username, msg) => {
       '%^^%⠀⠀⠀⠀⠀⠿⠀⠀⠈⠿⠿⠀⠈⠻⠶⠶⠾⠋⠀⠀⠀⠸⠇⠀⠀⠻⠶⠶⠿⠃');
 });
 
-HANDLERS.set('valorant', async (bot, username, msg) => {
-  if (!bot.valorantLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('valorant', async (bot, username, msg) => {
+  if (!Limiters.VALORANT.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$valorant is on cooldown');
   }
 
@@ -168,8 +137,8 @@ HANDLERS.set('valorant', async (bot, username, msg) => {
       '%^^%　　　     COMPLETE');
 });
 
-HANDLERS.set('osu', async (bot, username, msg) => {
-  if (!bot.osuLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('osu', async (bot, username, msg) => {
+  if (!Limiters.OSU.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$osu is on cooldown');
   }
 
@@ -182,8 +151,8 @@ HANDLERS.set('osu', async (bot, username, msg) => {
       '%^^%　　　     COMPLETE');
 });
 
-HANDLERS.set('among', async (bot, username, msg) => {
-  if (!bot.amongLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('among', async (bot, username, msg) => {
+  if (!Limiters.AMONG.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$among is on cooldown');
   }
 
@@ -207,8 +176,8 @@ HANDLERS.set('among', async (bot, username, msg) => {
       '%^^%⠀⠀⠀⠀⠀⠀⢿⣿⣦⣄⣀⣠⣴⣿⣿⠁⠀⠈⠻⣿⣿⣿⣿⡿⠏');
 });
 
-HANDLERS.set('xqck', async (bot, username, msg) => {
-  if (!bot.xqcKLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('xqck', async (bot, username, msg) => {
+  if (!Limiters.XQCK.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$xqcK is on cooldown');
   }
 
@@ -229,8 +198,8 @@ HANDLERS.set('xqck', async (bot, username, msg) => {
       '%^^%⣿⣿⣿⣦⡀⠀⠀⠉⠉⢉⣴⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿');
 });
 
-HANDLERS.set('pepejam', async (bot, username, msg) => {
-  if (!bot.pepejamLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('pepejam', async (bot, username, msg) => {
+  if (!Limiters.PEPE_JAM.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$pepejam is on cooldown');
   }
 
@@ -250,8 +219,9 @@ HANDLERS.set('pepejam', async (bot, username, msg) => {
       '%^^%⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣦');
 });
 
-HANDLERS.set('fuckyouchat', async (bot, username, msg) => {
-  if (!bot.fuckyouchatLimiter.tryRemoveTokens(1)) {
+// eslint-disable-next-line valid-jsdoc
+const /** @type {Handler} */ fuckYouChatHandler = async (bot, username, msg) => {
+  if (!Limiters.FUCK_YOU_CHAT.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$fuckyouchat is on cooldown');
   }
 
@@ -267,29 +237,13 @@ HANDLERS.set('fuckyouchat', async (bot, username, msg) => {
       '%^^%⣾⣿⠋⠙⠿⠗⠄⣿⣿⣀⣀⣿⣿⠄⠄⣸⣿⢿⣷⠄⠛⠛⣿⣿⠛⠛' +
       '%^^%⣿⣿⠄⠄⣀⠄⠄⣿⣿⠿⠿⣿⣿⠄⢠⣿⣏⣸⣿⡆⠄⠄⣿⣿' +
       '%^^%⠻⣿⣦⣴⣿⡟⠄⣿⣿⠄⠄⣿⣿⠄⣼⣿⠿⠿⢿⣿⡀⠄⣿⣿');
-});
+};
 
-HANDLERS.set('firetruckyouchat', async (bot, username, msg) => {
-  if (!bot.fuckyouchatLimiter.tryRemoveTokens(1)) {
-    return bot.sendPm(username, '$fuckyouchat is on cooldown');
-  }
+COMMANDS.set('fuckyouchat', fuckYouChatHandler);
+COMMANDS.set('firetruckyouchat', fuckYouChatHandler);
 
-  bot.sendChatMsg(
-      '%^^%⣿⡟⠛⠛⠃⢸⣿⡇⠄⠄⣿⡇⠄⣼⣿⠟⠻⣿⣆⠄⣿⣿⢠⣾⣿⠋' +
-      '%^^%⣿⣷⣶⣶⠄⢸⣿⡇⠄⠄⣿⡇⠄⣿⡏⠄⠄⠄⠄⠄⣿⣿⣿⣿⣇' +
-      '%^^%⣿⡇⠄⠄⠄⠘⣿⣧⣀⣰⣿⡇⠄⢿⣿⣀⣠⣿⡶⠄⣿⣿⠃⢹⣿⣆' +
-      '%^^%⠛⠃⠄⠄⠄⠄⠘⠛⠛⠛⠋⠄⠄⠈⠛⠛⠛⠛⠁⠄⠛⠛⠄⠄⠛⠛⠃' +
-      '%^^%⣤⡄⠄⠄⣤⣤⠄⢀⣠⣤⣄⡀⠄⢠⣤⡄⠄⠄⣤⣤' +
-      '%^^%⢻⣿⣄⣼⣿⠃⣰⣿⠟⠛⢿⣿⡄⢸⣿⡇⠄⠄⣿⣿' +
-      '%^^%⠄⠻⣿⡿⠁⠄⣿⣿⠄⠄⢸⣿⡇⢸⣿⡇⠄⠄⣿⣿' +
-      '%^^%⠄⠄⣿⡇⠄⠄⠹⣿⣦⣤⣼⣿⠃⠄⣿⣷⣤⣴⣿⡏' +
-      '%^^%⣾⣿⠋⠙⠿⠗⠄⣿⣿⣀⣀⣿⣿⠄⠄⣸⣿⢿⣷⠄⠛⠛⣿⣿⠛⠛' +
-      '%^^%⣿⣿⠄⠄⣀⠄⠄⣿⣿⠿⠿⣿⣿⠄⢠⣿⣏⣸⣿⡆⠄⠄⣿⣿' +
-      '%^^%⠻⣿⣦⣴⣿⡟⠄⣿⣿⠄⠄⣿⣿⠄⣼⣿⠿⠿⢿⣿⡀⠄⣿⣿');
-});
-
-HANDLERS.set('amonge', async (bot, username, msg) => {
-  if (!bot.amongeLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('amonge', async (bot, username, msg) => {
+  if (!Limiters.AMONG_E.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$amonge is on cooldown');
   }
 
@@ -307,8 +261,8 @@ HANDLERS.set('amonge', async (bot, username, msg) => {
       '%^^%⠄⢿⣿⣿⣿⣿⣿⣿⣿⡿⠃⢸⠿⠛⠉⣁⣠⣿⣿⣿⣿⣿⣿⠄⣼');
 });
 
-HANDLERS.set('gigachad', async (bot, username, msg) => {
-  if (!bot.gigachadLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('gigachad', async (bot, username, msg) => {
+  if (!Limiters.GIGACHAD.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$gigachad is on cooldown');
   }
 
@@ -323,17 +277,10 @@ HANDLERS.set('gigachad', async (bot, username, msg) => {
       '%^^%⣿⠋⢀⡌⢰⣿⡿⢿⡀⠀⠀⠀⠀⠀⠙⠿⣿⣿⣿⣿⣿⡇⠀⢸⣿⣿⣧' +
       '%^^%⣿⣷⢻⠄⠘⠛⠋⠛⠃⠀⠀⠀⠀⠀⢿⣧⠈⠉⠙⠛⠋⠀⠀⠀⣿⣿⣿' +
       '%^^%⣿⣧⠀⠈⢸⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠟⠀⠀⠀⠀⢀⢃⠀⠀⢸⣿⣿⣿');
-  // '%^^%⣿⡿⠀⠴⢗⣠⣤⣴⡶⠶⠖⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⡸⠀⣿⣿' +
-  // '%^^%⣿⣿⡀⢠⣾⣿⠏⠀⠠⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠛⠉⠀⣿⣿' +
-  // '%^^%⣿⣿⣧⠈⢹⡇⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣰⣿⣿' +
-  // '%^^%⣿⣿⣿⡄⠈⠃⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣠⣴⣾⣿⣿' +
-  // '%^^%⣿⣿⣿⣧⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣠⣾⣿⣿⣿⣿⣿⣿' +
-  // '%^^%⣿⣿⣿⣷⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣴⣿⣿⣿⣿⣿⣿⣿⣿⣿' +
-  // '%^^%⣿⣿⣿⣿⣦⣄⣀⣀⣀⣀⠀⠀⠀⠀⠘⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿');
 });
 
-HANDLERS.set('kkona', async (bot, username, msg) => {
-  if (!bot.kkonaLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('kkona', async (bot, username, msg) => {
+  if (!Limiters.KKONA.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$kkona is on cooldown');
   }
 
@@ -349,8 +296,8 @@ HANDLERS.set('kkona', async (bot, username, msg) => {
       `%^^%⠈⠛⠛⠛⠓⠒⠚⠛⠛⠛⠛⠛⠋⠙⠛⠛⠛⠛⠟⠿⠷⠶⠶⠾⠿⢿⠿ Don't mind me, just a feller out on the farm.`);
 });
 
-HANDLERS.set('ayaya', async (bot, username, msg) => {
-  if (!bot.ayayaLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('ayaya', async (bot, username, msg) => {
+  if (!Limiters.AYAYA.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$ayaya is on cooldown');
   }
 
@@ -365,11 +312,10 @@ HANDLERS.set('ayaya', async (bot, username, msg) => {
       '%^^%⠷⡹⣿⠋⣉⣠⣤⣶⣶⣿⣿⣿⣿⣿⣿⡿⠿⢿⣿⣿⣿⣿⣿⣷⢹⣿⣿' +
       '%^^%⡝⣿⡞⣿⣿⣿⣿⣿⣿⣿⣿⡟⠋⠁⣠⣤⣤⣦⣽⣿⣿⣿⡿⠋⠘⣿⣿' +
       '%^^%⣿⡹⣿⡼⣿⣿⣿⣿⣿⣿⣿⣧⡰⣿⣿⣿⣿⣿⣹⡿⠟⠉⡀⠄⠄⢿⣿');
-  // '%^^%⣿⣿⣽⣿⣼⣛⠿⠿⣿⣿⣿⣿⣿⣯⣿⠿⢟⣻⡽⢚⣤⡞⠄⠄⠄⢸⣿');
 });
 
-HANDLERS.set('flip', async (bot, username, msg) => {
-  if (!bot.flipLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('flip', async (bot, username, msg) => {
+  if (!Limiters.FLIP.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$flip is on cooldown');
   }
 
@@ -382,8 +328,8 @@ HANDLERS.set('flip', async (bot, username, msg) => {
       '%^^%(╯°□°)╯︵ ┻━┻ NEVER!!!');
 });
 
-HANDLERS.set('kappa', async (bot, username, msg) => {
-  if (!bot.kappaLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('kappa', async (bot, username, msg) => {
+  if (!Limiters.KAPPA.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$kappa is on cooldown');
   }
 
@@ -401,11 +347,10 @@ HANDLERS.set('kappa', async (bot, username, msg) => {
       '%^^%░░▀░░░░░░▒▄▄▒▄▄▄▒▒█░' +
       '%^^%░░░▀▄▄▒▒░░░░▀▀▒▒▄▀░░' +
       '%^^%░░░░░▀█▄▒▒░░░░▒▄▀░░░');
-  // '%^^%░░░░░░░░▀▀█▄▄▄▄▀░░░');
 });
 
-HANDLERS.set('pagchomp', async (bot, username, msg) => {
-  if (!bot.pagchompLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('pagchomp', async (bot, username, msg) => {
+  if (!Limiters.PAG_CHOMP.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$pagchomp is on cooldown');
   }
 
@@ -421,14 +366,10 @@ HANDLERS.set('pagchomp', async (bot, username, msg) => {
       '%^^%⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣥⣄⣀⣀⠀⠀⠀⠀⠀⢰⣾⣿⣿⠏' +
       '%^^%⠀⠀⠉⣩⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣶⣤⣜⡻⠋' +
       '%^^%⣿⣾⣖⣻⣿⣿⡿⣿⣿⣿⣿⠿⠿⠟⠛⠛⠛⠋⠉⠉⢉⡽⠃');
-  // '%^^%⣿⣿⡉⠉⠉⠛⠛⠛⠉⠁⠀⠀⠀⠀⠀⠀⠀⠀⡤⠚⠉' +
-  // '%^^%⣿⣿⣿⣿⣿⠉⠛⢶⣶⣄⠀⠀⠀⠀⠀⠀⠀⠀⡇' +
-  // '%^^%⣿⣿⣿⠿⠟⠃⠀⠀⠀⠈⠲⣴⣦⣤⣤⣤⣶⡾⠁' +
-  // '%^^%⢿⣿⣿⡀⠀⠀⠀⠀⠀⠀⠀⠀⠉⠛⠛⠛⠉');
 });
 
-HANDLERS.set('bedge', async (bot, username, msg) => {
-  if (!bot.bedgeLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('bedge', async (bot, username, msg) => {
+  if (!Limiters.BEDGE.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$bedge is on cooldown');
   }
 
@@ -450,8 +391,8 @@ HANDLERS.set('bedge', async (bot, username, msg) => {
       '%^^%⣿⡯⠓⠁⠀⠀⠀⠀⠀⠀⠀⠀⠈⠁⠀⠀⠀⠀⠀⣀⡤⠒⠋⠁');
 });
 
-HANDLERS.set('starege', async (bot, username, msg) => {
-  if (!bot.staregeLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('starege', async (bot, username, msg) => {
+  if (!Limiters.STAREGE.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$starege is on cooldown');
   }
 
@@ -471,8 +412,8 @@ HANDLERS.set('starege', async (bot, username, msg) => {
       '%^^%⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿');
 });
 
-HANDLERS.set('poggers', async (bot, username, msg) => {
-  if (!bot.poggersLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('poggers', async (bot, username, msg) => {
+  if (!Limiters.POGGERS.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$poggers is on cooldown');
   }
 
@@ -488,14 +429,10 @@ HANDLERS.set('poggers', async (bot, username, msg) => {
       '%^^%⠄⠙⣿⡄⠄⠄⢀⡤⠬⢭⣝⣒⢂⠭⣉⠻⠿⣷⣶⣦⣭⡛⣿⣿⣿' +
       '%^^%⠄⠄⠸⣿⡇⠄⠸⣎⣁⣾⠿⠉⢀⠇⣸⣿⣿⢆⡉⠹⣿⣿⢸⣿⣿' +
       '%^^%⠄⠄⠄⣿⡇⠄⢀⡶⠶⠶⠾⠿⠮⠭⠭⢭⣥⣿⣿⣷⢸⣿⢸⣿⣿');
-  // '%^^%⠄⠄⠄⣿⡇⠄⠈⣷⠄⠄⠄⣭⣙⣹⢙⣰⡎⣿⢏⣡⣾⢏⣾⣿⣿' +
-  // '%^^%⠄⠄⢰⣿⡇⠄⠄⢿⠄⠄⠈⣿⠉⠉⣻⣿⡷⣰⣿⡿⣡⣿⣿⣿⣿' +
-  // '%^^%⠄⠄⢸⣿⡇⠄⠄⠘⠿⠤⠤⠿⠿⠿⢤⣤⣤⡿⣃⣸⣿⣿⣿⣿⣿' +
-  // '%^^%⠄⠄⠘⢿⣷⣤⣄⣀⣀⣀⣀⣀⣠⣴⣾⡿⢋⣼⣿⣿⣿⣿⣿⣿⣿');
 });
 
-HANDLERS.set('peeposad', async (bot, username, msg) => {
-  if (!bot.peeposadLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('peeposad', async (bot, username, msg) => {
+  if (!Limiters.PEEPO_SAD.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$peeposad is on cooldown');
   }
 
@@ -512,8 +449,8 @@ HANDLERS.set('peeposad', async (bot, username, msg) => {
       '%^^%⢣⣿⣿⣿⣿⣿⣿⣿⣶⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⡝');
 });
 
-HANDLERS.set('monkas', async (bot, username, msg) => {
-  if (!bot.monkasLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('monkas', async (bot, username, msg) => {
+  if (!Limiters.MONKA_S.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$monkas is on cooldown');
   }
 
@@ -534,8 +471,8 @@ HANDLERS.set('monkas', async (bot, username, msg) => {
       '%^^%⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠉⢻⣿⣿⣿');
 });
 
-HANDLERS.set('hypers', async (bot, username, msg) => {
-  if (!bot.hypersLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('hypers', async (bot, username, msg) => {
+  if (!Limiters.HYPERS.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$hypers is on cooldown');
   }
 
@@ -550,15 +487,10 @@ HANDLERS.set('hypers', async (bot, username, msg) => {
       '%^^%⡇⠀⠀⠀⠀⠀⣀⣀⣀⣉⣉⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣀⣠⣤⣴⡿⢧⡇⠀⣾' +
       '%^^%⡇⠀⠀⠀⠀⠸⣿⣯⣙⣛⡛⠿⠿⠿⠿⠿⠿⠿⠟⢛⣛⣋⣭⣥⡖⡿⠁⠀⢠⣿' +
       '%^^%⣧⠀⠀⠀⠀⠀⠀⠀⠈⠉⠉⠛⠛⠛⠛⠛⠛⠛⠛⠛⠉⠉⠉⠀⠀⠀⠀⢀⣾⣿');
-  // '%^^%⣿⡀⠀⠀⠐⢦⡀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⣀⣠⣤⡶⣠⣴⣿⣿⣿' +
-  // '%^^%⣿⣷⠀⠀⠀⠀⣿⣿⣶⣤⣤⣤⣤⣤⣤⣶⣶⣶⣶⣿⣿⣿⣿⣿⣷⣌⠻⣿⣿⣿' +
-  // '%^^%⣿⡏⣾⣦⣄⣰⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣷⢹⣿⣿' +
-  // '%^^%⣿⣇⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⢟⣼⣿⣿' +
-  // '%^^%⣿⣿⣷⣶⣯⣭⣭⣽⣟⣛⣛⣛⣛⣿⣿⣛⣛⣛⣛⣛⣫⣭⣭⣿⣶⣾⣿⣿⣿⣿');
 });
 
-HANDLERS.set('yep', async (bot, username, msg) => {
-  if (!bot.yepLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('yep', async (bot, username, msg) => {
+  if (!Limiters.YEP.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$yep is on cooldown');
   }
 
@@ -576,8 +508,8 @@ HANDLERS.set('yep', async (bot, username, msg) => {
       '%^^%⠹⣿⣄⣠⣶⡆⠄⢿⣷⣀⣠⣿⡟⠄⢻⣷⣄⣠⣶⠆⠄⣿⡏⠈⢿');
 });
 
-HANDLERS.set('monkaw', async (bot, username, msg) => {
-  if (!bot.monkawLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('monkaw', async (bot, username, msg) => {
+  if (!Limiters.MONKA_W.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$monkaw is on cooldown');
   }
 
@@ -593,14 +525,10 @@ HANDLERS.set('monkaw', async (bot, username, msg) => {
       '%^^%⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠈⠥⠀⠀⠀⠀⠈⢷⡀' +
       '%^^%⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢻' +
       '%^^%⣀⣀');
-  // '%^^%⠛⠛⠹⠿⠻⠶⢶⣦⣤⣤⣀⣀⣀⣀⣀⣀⣀⣀⢀⣀⣀⡀⣀⣀⣀⣠⣴⡶⠿' +
-  // '%^^%⣶⣦⣤⣴⣄⣀⡨⠀⡁⠈⠉⠉⠛⠻⠛⠛⠛⠿⠿⠿⠿⠟⠛⠛⠋⠉⢈⣨⣤' +
-  // '%^^%⠀⠀⠉⠉⠉⠉⠋⠛⠳⢷⠶⣾⣯⣦⣤⣤⣤⣤⣤⣤⣤⣴⣶⡾⣧⣞⠛⣻⠁' +
-  // '%^^%⠿⠷⠶⢿⣦⣤⣦⣰⣄⣄⢀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⠀⢀⣀⣰⣼⠟');
 });
 
-HANDLERS.set('pepelaugh', async (bot, username, msg) => {
-  if (!bot.pepelaughLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('pepelaugh', async (bot, username, msg) => {
+  if (!Limiters.PEPE_LAUGH.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$pepelaugh is on cooldown');
   }
 
@@ -615,13 +543,10 @@ HANDLERS.set('pepelaugh', async (bot, username, msg) => {
       '%^^%⠄⠄⠄⠄⠄⠄⠄⢠⠄⡖⢉⠥⢤⠐⢲⠒⢲⠒⢲⠒⠲⡒⠒⡖⢲⠂⠄⢀⣿' +
       '%^^%⠄⠄⠄⠄⠄⠄⠄⠈⢆⡑⢄⠳⢾⠒⢺⠒⢺⠒⠚⡖⠄⡏⠉⣞⠞⠁⣠⣾' +
       '%^^%⠄⠄⠄⠄⠄⢆⠄⠄⠄⠈⠢⠉⠢⠍⣘⣒⣚⣒⣚⣒⣒⣉⠡⠤⣔⣾⣿');
-  // '%^^%⣤⠄⣀⠄⠄⠄⠈⠁⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⢀⣤⣾⣿' +
-  // '%^^%⠄⠉⠐⠢⠭⠄⢀⣒⣒⡒⠄⠄⠄⠄⠄⠄⣀⡠⠶⢶⣿⣿' +
-  // '%^^%⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠄⠈⠁⠈⠄⠄⠄⠄⠄⠄⠈⠻⣿');
 });
 
-HANDLERS.set('pepega', async (bot, username, msg) => {
-  if (!bot.pepegaLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('pepega', async (bot, username, msg) => {
+  if (!Limiters.PEPEGA.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$pepega is on cooldown Pepega');
   }
 
@@ -639,8 +564,8 @@ HANDLERS.set('pepega', async (bot, username, msg) => {
       '%^^%⠁⠁⠁⠉⠻⢿⣿⣿⣷⣦⣬⣍⣓⡒⠒⣒⣂⣠⡬⠽⠓⠂⠁');
 });
 
-HANDLERS.set('uwu', async (bot, username, msg) => {
-  if (!bot.uwuLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('uwu', async (bot, username, msg) => {
+  if (!Limiters.UWU.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$uwu is on cooldown');
   }
 
@@ -655,12 +580,10 @@ HANDLERS.set('uwu', async (bot, username, msg) => {
       '%^^%⢕⠅⣐⢕⢕⢕⢕⢕⣿⣿⡄⠛⢀⣦⠈⠛⢁⣼⣿⢗⢕⢕⢕⢕⢕⢕⡏' +
       '%^^%⢕⠅⢓⣕⣕⣕⣕⣵⣿⣿⣿⣾⣿⣿⣿⣿⣿⣿⣿⣷⣕⢕⢕⢕⢕⡵' +
       '%^^%⢕⠃⡈⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⢃');
-  // '%^^%⢕⠄⢱⣄⠛⢿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿⠿⢁' +
-  // '%^^%⣦⡀⣿⣿⣷⣶⣬⣍⣛⣛⣛⡛⠿⠿⠿⠛⠛⢛⣛⣉⣭⣤⣂⢜');
 });
 
-HANDLERS.set('tf', async (bot, username, msg) => {
-  if (!bot.tfLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('tf', async (bot, username, msg) => {
+  if (!Limiters.TF.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$tf is on cooldown');
   }
 
@@ -682,8 +605,8 @@ HANDLERS.set('tf', async (bot, username, msg) => {
       '%^^%░░░░░░░░░░░░░░▀▄▄▄▄▄░░░░░░░░█░░');
 });
 
-HANDLERS.set('gachi', async (bot, username, msg) => {
-  if (!bot.gachiLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('gachi', async (bot, username, msg) => {
+  if (!Limiters.GACHI.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$gachi is on cooldown');
   }
 
@@ -700,11 +623,10 @@ HANDLERS.set('gachi', async (bot, username, msg) => {
       '%^^%░░▓█▒▒▄▒▒▒▒▒▒▒░░▒▒▒▒▒▒▓' +
       '%^^%░░░▓▓▓▓▒▒▒▒▒▒▒▒░░░▒▒▒▓▓' +
       '%^^%░░░░░▓▓▒░░▒▒▒▒▒▒▒▒▒▒▒▓▓');
-  // '%^^%░░░░░░▓▒▒░░░░▒▒▒▒▒▒▒▓▓');
 });
 
-HANDLERS.set('forgor', async (bot, username, msg) => {
-  if (!bot.forgorLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('forgor', async (bot, username, msg) => {
+  if (!Limiters.FORGOR.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$forgor is on cooldown');
   }
 
@@ -725,9 +647,9 @@ HANDLERS.set('forgor', async (bot, username, msg) => {
       '%^^%⠀⠀⠀/ \\');
 });
 
-HANDLERS.set('fortnite', async (bot, username, msg) => {
+COMMANDS.set('fortnite', async (bot, username, msg) => {
   if (!(await bot.checkPermission(username, Rank.MOD, 'I'))) {
-    if (!bot.fortniteLimiter.tryRemoveTokens(1)) {
+    if (!Limiters.FORTNITE.tryRemoveTokens(1)) {
       return bot.sendPm(username, '$fortnite is on cooldown');
     }
   }
@@ -1156,8 +1078,8 @@ HANDLERS.set('fortnite', async (bot, username, msg) => {
       '%^^%⠀⠀⠀⠀⣶⠿⠿⠛');
 });
 
-HANDLERS.set('nomods', async (bot, username, msg) => {
-  if (!bot.nomodsLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('nomods', async (bot, username, msg) => {
+  if (!Limiters.NO_MODS.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$nomods is on cooldown');
   }
 
@@ -1178,8 +1100,8 @@ HANDLERS.set('nomods', async (bot, username, msg) => {
       '%^^%⣿⣷⣶⣿⣶⣾⣷⣶⣿⣶⣶⣶⣿⣷⣶⣶⣶⣾⣶⣶⣶⣾');
 });
 
-HANDLERS.set('moderators', async (bot, username, msg) => {
-  if (!bot.modscmdLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('moderators', async (bot, username, msg) => {
+  if (!Limiters.MODS_SMELL_BAD.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$mods is on cooldown');
   }
 
@@ -1198,8 +1120,8 @@ HANDLERS.set('moderators', async (bot, username, msg) => {
       '%^^%⣿⣿⣿⣿⣿⣿⣿⣿⣷⣶⣶⣾⣿⣿⣿⣿⣿⣿⣿⣿⣿⣿');
 });
 
-HANDLERS.set('onion', async (bot, username, msg) => {
-  if (!bot.onionLimiter.tryRemoveTokens(1)) {
+COMMANDS.set('onion', async (bot, username, msg) => {
+  if (!Limiters.ONION.tryRemoveTokens(1)) {
     return bot.sendPm(username, '$onion is on cooldown');
   }
 
@@ -1222,61 +1144,4 @@ HANDLERS.set('onion', async (bot, username, msg) => {
       '%^^%░░░░█▄░░░░░░░░▄█' +
       '%^^%░░░░░░░▀██████▀' +
       '%^^%░░░░░░░░░░▀▀');
-});
-
-HANDLERS.set('gn', (bot, username, msg) => {
-  bot.sendChatMsg(`FeelsOkayMan <3 gn ${username}`);
-});
-
-/** Function that generates a random number based on a normal distribution. */
-const normal = random.normal(/** mu= */ 100, /** sigma= */ 15);
-
-HANDLERS.set('iq', (bot, username, msg) => {
-  const target = msg.split(' ')[0] !== '' ? msg.split(' ')[0] : username;
-  const iq = Math.round(normal());
-
-  let /** @type {string} */ emote;
-  if (iq > 115) {
-    emote = ', ah yes 6Head';
-  } else if (iq < 85) {
-    emote = 'Pepege';
-  } else {
-    emote = ', average ForsenLookingAtYou';
-  }
-
-  bot.sendChatMsg(`${target}'s IQ is ${iq} ${emote}`);
-});
-
-HANDLERS.set('pyramid', async (bot, username, msg) => {
-  if (!msg) {
-    return;
-  }
-
-  if (!bot.pyramidLimiter.tryRemoveTokens(1)) {
-    return bot.sendPm(username, '$pyramid is on cooldown');
-  }
-
-  // Send an invisible character first so all of the pyramid lines are aligned (second and
-  // subsequent message in a row are left-aligned in chat)
-  bot.sendChatMsg('⠀');
-
-  const word = msg.split(' ')[0];
-  bot.sendChatMsg(`⠀ ${word}`);
-  bot.sendChatMsg(`⠀ ${word} ${word}`);
-  bot.sendChatMsg(`⠀ ${word} ${word} ${word}`);
-  bot.sendChatMsg(`⠀ ${word} ${word} ${word} ${word}`);
-  bot.sendChatMsg(`⠀ ${word} ${word} ${word}`);
-  bot.sendChatMsg(`⠀ ${word} ${word}`);
-  bot.sendChatMsg(`⠀ ${word}`);
-});
-
-HANDLERS.set('tuck', (bot, username, msg) => {
-  const target = msg.split(' ')[0];
-  if (target === '') {
-    bot.sendChatMsg(
-        `${username}, who do you want to tuck in? FeelsOkayMan ` +
-        'Example: $tuck MrDestructoidCyDJ');
-  }
-
-  bot.sendChatMsg(`Bedge ${username} tucks ${target} into bed.`);
 });
