+++ lib/cytubebot.js
@@ -7,7 +7,7 @@ import socketIoClient from 'socket.io-client';
 
 import {lookupSocketUrl} from './apiclient.js';
 import {addHandlers} from './bothandlers.js';
-import {handle as chatHandle} from './chatcommands.js';
+import {handleChatMessage} from './commands/handle.js';
 import {RESTART_TIMES_FILE_NAME} from './constants.js';
 import {Cookies} from './cookie.js';
 import {Database} from './database.js';
@@ -230,81 +230,11 @@ export class CytubeBot {
     /** @type {RateLimiter} */
     this.showemoteLimiter = new RateLimiter({tokensPerInterval: 1, interval: 15 * 1_000});
     /** @type {RateLimiter} */
-    this.forsenLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.potfriendLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.trolldespairLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.deezLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.valorantLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.amongLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.xqcKLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.pepejamLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.fuckyouchatLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.amongeLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.gigachadLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.kkonaLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.ayayaLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.flipLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.kappaLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.pagchompLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.bedgeLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.staregeLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.poggersLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.peeposadLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.monkasLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.hypersLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.yepLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.monkawLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.pepelaughLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.pepegaLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.uwuLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.tfLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.gachiLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.forgorLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.fortniteLimiter = new RateLimiter({tokensPerInterval: 1, interval: 900 * 1_000});
-    /** @type {RateLimiter} */
-    this.nomodsLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.modscmdLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
-    this.onionLimiter = new RateLimiter({tokensPerInterval: 1, interval: 30 * 1_000});
-    /** @type {RateLimiter} */
     this.hereGlobalLimiter = new RateLimiter({tokensPerInterval: 1, interval: 300 * 1_000});
     /** @type {RateLimiter} */
     this.modscmdGlobalLimiter = new RateLimiter({tokensPerInterval: 1, interval: 20 * 1_000});
     /** @type {RateLimiter} */
     this.staffcmdLimiter = new RateLimiter({tokensPerInterval: 1, interval: 60 * 1_000});
-    /** @type {RateLimiter} */
-    this.osuLimiter = new RateLimiter({tokensPerInterval: 1, interval: 60 * 1_000});
     /** @type {number} */
     this.timeSinceLastAnagram = 0;
     /** @type {number} */
@@ -885,7 +815,7 @@ export class CytubeBot {
         (username.toLowerCase() !== this.username.toLowerCase()) && this.loggedIn && allowed;
 
     if (handleCommand) {
-      chatHandle(this, username, msg);
+      handleChatMessage(this, username, msg);
       return;
     }
